[기능 수정] 크루 상세 컴포넌트 세분화
