+++ frontend/src/components/Crew/CrewDetail/CrewDetail.tsx
@@ -15,6 +15,9 @@ import DetailBaseBorder from '../../../common/components/DetailBaseBorder';
 import PreviousPageButton from '../../../common/components/PreviousPageButton';
 import NextPageButton from '../../../common/components/NextPageButton';
 import LeaveCrewButton from './LeaveCrewButton';
+import CrewNormalInformation from './CrewNormalInformation';
+import CrewSignUpRequestButton from './CrewSignUpRequestButton';
+import CrewDetailInformation from './CrewDetailInformation';
 
 interface MatchParam {
   id: string;
@@ -27,40 +30,21 @@ const CrewDetail: React.FC<RouteComponentProps<MatchParam>> = ({ match }) => {
   const {
     crewName,
     crewLeaderId,
-    crewRegion,
-    explanation,
-    openChat,
-    crewUserCount,
     crewRequestFetch,
-    requestUsers,
-    crewRequested,
     userId,
     userNickName,
     isCrewLeader,
     userCrewName,
-    token,
   } = useSelector((state) => ({
     crewName: state.crew.crewName,
     crewLeaderId: state.crew.crewLeaderId,
-    crewRegion: state.crew.crewRegion,
-    explanation: state.crew.explanation,
-    openChat: state.crew.openChat,
-    crewUserCount: state.crew.userDtos.length,
     crewRequestFetch: state.crew.crewRequestFetch,
-    requestUsers: state.crew.requestUsers,
-    crewRequested: state.crew.crewRequested,
     userId: state.signIn.userData.id,
     userNickName: state.signIn.userData.nickName,
     isCrewLeader: state.signIn.userData.crewLeader,
     userCrewName: state.signIn.userData.crewName,
-    token: state.signIn.token,
   }));
   const { getUserData } = useLocalStroeageData();
-
-  const normalCategory = [
-    { icon: BsPeopleFill, title: '크루 인원', description: crewUserCount },
-    { icon: GiPositionMarker, title: '크루 지역', description: crewRegion },
-  ];
 
   useEffect(() => {
     new CrewService()
@@ -111,10 +95,6 @@ const CrewDetail: React.FC<RouteComponentProps<MatchParam>> = ({ match }) => {
     getUserData();
   }, []);
 
-  const signUpRequestCrew = () => {
-    dispatch(crewActions.signUpRequestCrew({ crewName, token }));
-  };
-
   return (
     <DetailBaseBorder>
       <div className="flex items-center justify-between">
@@ -136,24 +116,7 @@ const CrewDetail: React.FC<RouteComponentProps<MatchParam>> = ({ match }) => {
         )}
       </div>
       <div className="w-full mx-auto my-0 py-5 flex flex-col flex-wrap justify-center items-center space-y-5">
-        <div className="w-full flex justify-center items-center">
-          <img
-            src=""
-            alt=""
-            className="w-48 rounded-full border-4 border-purple "
-          />
-        </div>
-        <div className="text-2xl">{crewName}</div>
-        <div className="text-lg">
-          <span>{explanation}</span>
-        </div>
-        <div className="text-lg">
-          <span>
-            <a href={openChat} target="_blank" rel="noreferrer">
-              오픈 채팅 : {openChat}
-            </a>
-          </span>
-        </div>
+        <CrewDetailInformation />
         {userCrewName === crewName && +userId !== crewLeaderId && (
           <div className="w-full ml-3 lg:ml-10">
             <LeaveCrewButton />
@@ -161,33 +124,8 @@ const CrewDetail: React.FC<RouteComponentProps<MatchParam>> = ({ match }) => {
         )}
       </div>
       <div className="space-y-5">
-        <div className="w-20 md:w-44 lg:w-52 pl-5 md:pl-0 py-4 flex flex-grow justify-start items-start">
-          {+userId !== crewLeaderId &&
-            userCrewName !== crewName &&
-            !isCrewLeader && (
-              <Button
-                auto
-                color="#8b8bf5"
-                onClick={signUpRequestCrew}
-                disabled={crewRequested}
-              >
-                {crewRequested ? '요청됨' : '가입요청 보내기'}
-              </Button>
-            )}
-        </div>
-        <span className="block pl-5 md:pl-0 text-lg">기본정보</span>
-        <div className="w-full grid grid-cols-1 sm:grid-cols-2 xl:grid-cols-3 mx-auto gap-5">
-          {normalCategory.map((category) => (
-            <div key={v4()} className="flex justify-center">
-              <CrewWidget
-                Icon={category.icon}
-                widgetTitle={category.title}
-                widgetDescription={category.description}
-                iconColor="#8b8bf5"
-              />
-            </div>
-          ))}
-        </div>
+        <CrewSignUpRequestButton />
+        <CrewNormalInformation />
       </div>
     </DetailBaseBorder>
   );
